Unified `ExtentId` type (#1310)
Across the codebase, we refer to extent IDs as either `u32`, `u64`, or
`usize`.

This PR switches to an `ExtentId(pub u32)` type, and uses it everywhere.
It is a **breaking change** to the message protocol, because extent IDs
were previously encoded as `u64`'s in messages.

For consistency, it also changes `extent_count: u64` to `extent_count:
u32` in various places.

This is Episode 5 of a long-running series; previously:

- https://github.com/oxidecomputer/crucible/pull/919
- https://github.com/oxidecomputer/crucible/pull/925
- https://github.com/oxidecomputer/crucible/pull/928
- https://github.com/oxidecomputer/crucible/pull/1076

+++ agent/src/model.rs
@@ -36,7 +36,7 @@ pub struct Region {
     // Creation parameters
     pub block_size: u64,
     pub extent_size: u64,
-    pub extent_count: u64,
+    pub extent_count: u32,
     pub encrypted: bool,
 
     // Run-time parameters
@@ -130,7 +130,7 @@ pub struct CreateRegion {
 
     pub block_size: u64,
     pub extent_size: u64,
-    pub extent_count: u64,
+    pub extent_count: u32,
     pub encrypted: bool,
 
     pub cert_pem: Option<String>,
